shifted code from person to mynetwork
